requirements.txt
scrapy
selenium
beautifulsoup4
requests
pandas
openpyxl
supabase
python-dotenv
sentence-transformers
scikit-learn
numpy
groq
python-docx
streamlit